% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-generic.R
\name{extract}
\alias{extract}
\title{Generic for \code{extract} method.}
\usage{
extract(object, ...)
}
\arguments{
\item{object}{The object on which the generic operates}

\item{\dots}{Arguments passed to specific methods}
}
\description{
Generic for \code{extract} method.
}

