% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{extract}
\alias{extract}
\alias{missingpreds}
\title{Generics for \code{extract} and \code{missingpreds} methods}
\usage{
extract(object, ...)
missingpreds(object, ...)
}
\arguments{
\item{object}{The object on which the generic operates}

\item{\dots}{Arguments passed to specific methods}
}
\description{
Generics for \code{extract} and \code{missingpreds} methods
}

