% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{glmnetLRC-package}
\alias{glmnetLRC-package}
\title{Elastic net logistic regression classification (LRC) with an arbitrary loss function}
\description{
\pkg{glmnetLRC} extends the \pkg{glmnet}
package for training elastic net LRCs with a user-specified discrete loss function.
}
\details{
There are three functions you'll need (along with their associated methods):
\code{\link{lossMatrix}}, \code{\link{glmnetLRC}} and \code{\link{predict.glmnetLRC}}.
}
\author{
Landon Sego, Alex Venzin
}

