% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnetLRC_fit.R
\docType{data}
\name{glmnetLRC_fit}
\alias{glmnetLRC_fit}
\title{A glmnetLRC model fit object}
\format{A glmnetLRC object returned by \code{\link{glmnetLRC}}}
\description{
This object is returned by the call
\code{glmnetLRC_fit <- glmnetLRC(response, predictors, lossMat = lM, estimateLoss = TRUE,
nJobs = parallel::detectCores())} in the example of \code{\link{glmnetLRC}}. It is preserved
here in the package because it is time consuming to generate.
}
\examples{
# Load fitted glmnetLRC model
data(glmnetLRC_fit)
}

