% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glmnetLRClicense.R
\name{glmnetLRClicense}
\alias{glmnetLRClicense}
\title{Open text version of public domain license}
\usage{
glmnetLRClicense()
}
\description{
Open text version of public domain license
}
\author{
Alexander Venzin
}

