% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrc-package.R
\docType{package}
\name{lrc-package}
\alias{lrc-package}
\title{Elastic net logistic regression classification (LRC) with an arbitrary loss function}
\description{
Elastic net logistic regression classification (LRC) with an arbitrary loss function
}
\details{
\pkg{lrc} extends the \pkg{glmnet} 
package for training elastic net LRCs with a user-specified
discrete loss function.

There are a handful of functions you'll need (along with their associated methods):
\code{\link{lossMatrix}}, \code{\link{LRCglmnet}}, 
and \code{\link{predict.LRCglmnet}}
}
\author{
Landon Sego, Alex Venzin, John Ramey
}

