% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LRCglmnet_fit.R
\docType{data}
\name{LRCglmnet_fit}
\alias{LRCglmnet_fit}
\title{A LRCglmnet model fit object}
\format{A LRCglmnet object returned by \code{\link{LRCglmnet}}}
\description{
This object is returned by the call
\code{LRCglmnet_fit <- LRCglmnet(response, predictors, lM, cores = max(1, detectCores() - 1))}
in the example of \code{\link{LRCglmnet}}. It is preserved here in the package
because it is time consuming to generate.
}
\examples{
# Load fitted LRCglmnet model
data(LRCglmnet_fit)

# Show the optimal parameter values
print(LRCglmnet_fit)

# Show the plot of all the optimal parameter values for each cross validation replicate
plot(LRCglmnet_fit)

# Load the new observations
data(testdata)

# Use the trained model to make predictions about
# new observations for the response variable.
new <- predict(LRCglmnet_fit, testdata, truthCol = "Curated_Quality", keepCols = 1:2)
head(new)

# Now summarize the performance of the model
summary(new)

# If predictions are made without the an indication of the ground truth,
# the summary is simpler:
summary(predict(LRCglmnet_fit, testdata))
}

