% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.permuteTest.R
\name{plot.permuteTest}
\alias{plot.permuteTest}
\title{Plot the null distribution of the permutation test}
\usage{
\method{plot}{permuteTest}(x, ...)
}
\arguments{
\item{x}{an object of class \code{permuteTest} returned by \code{\link{permuteTest.glmnetLRC}}.}

\item{\dots}{Arguments passed to \code{\link{hist}}}
}
\value{
Produces a histogram of the null distribution of the loss, with a vertical red
line indicating the mean expected loss.
}
\description{
Plot the null distribution of the permutation test
}
\seealso{
See \code{\link{glmnetLRC}} for an example.
}
\author{
Landon Sego
}
