% Generated by roxygen2 (4.0.0): do not edit by hand
\name{predict.glmnetLRC}
\alias{predict.glmnetLRC}
\alias{summary.glmnetLRCpred}
\title{Predict (or classify) new data using a fitted glmnet logistic regression classifier}
\usage{
\method{predict}{glmnetLRC}(glmnetLRCobject, newdata, truthCol = NULL,
  keepCols = NULL)

\method{summary}{glmnetLRCpred}(glmnetLRCpredObject)
}
\arguments{
\item{glmnetLRCobject}{An object of class \code{glmnetLRC}, returned by
\code{\link{glmnetLRC}},
which contains the optimally-trained elastic net logistic regression classifier}

\item{newdata}{A dataframe or matrix containing the new set of observations to
be predicted. \code{newdata} must contain all of the column names that were used
to fit the elastic net logistic regression classifier.}

\item{truthCol}{The column number or column name in \code{newdata} that contains the
true labels. Optional.}

\item{keepCols}{A numeric vector of column numbers (or a character vector of
column names) that will be 'kept' and returned with the predictions. Optional.}

\item{glmnetLRCpredObject}{an object of class \code{glmnetLRCpred} returned by
\code{predict.glmnetLRC}.}
}
\value{
\describe{
\item{\code{predict.glmnetLRC}}{
An object of class \code{glmnetLRCpred} (which inherits
from \code{data.frame}
that contains the predicted class for each observation.  The columns indicated
by \code{truthCol} and \code{keepCols} are included if they were requested.}

\item{\code{summary.glmnetLRCpred}}{

If \code{truthCol} was provided in the call to
\code{predict.glmnetLRC}, a
\code{data.frame} is returned with the sensitivity, specificity, false negative rate,
false positive rate, and accuracy for the class designated by the second level of
the \code{truthLabels} argument in \code{\link{glmnetLRC}}.

If \code{truthCol = NULL} in the call to \code{predict.glmnetLRC}, a tabulation
of the number of predictions for each class is returned.
}
}
}
\description{
Predict (or classify) new data using a fitted glmnet logistic regression classifier
}
\author{
Landon Sego
}
\seealso{
See \code{\link{glmnetLRC}} for examples
}

