% Generated by roxygen2 (4.0.1): do not edit by hand
\name{predict.LRCbestsubsets}
\alias{predict.LRCbestsubsets}
\title{Predict (or classify) new data using a fitted best subsets logistic regression
classifier}
\usage{
\method{predict}{LRCbestsubsets}(LRCbestsubsets_object, newdata,
  truthCol = NULL, keepCols = NULL)
}
\arguments{
\item{LRCbestsubsets_object}{An object of class \code{LRCbestsubsets}, returned by
\code{\link{LRCbestsubsets}},
which contains the best subsets logistic regression classifier with optimal threshold
\eqn{\tau}.}

\item{newdata}{A dataframe containing the new set of observations to
be predicted, as well as an optional column of true labels.
\code{newdata} must contain all of the column names that were used
to fit best subsets logistic regression classifier.}

\item{truthCol}{The column number or column name in \code{newdata} that contains the
true labels. Optional.}

\item{keepCols}{A numeric vector of column numbers (or a character vector of
column names) that will be 'kept' and returned with the predictions. Optional.}
}
\value{
\item{\code{predict.LRCbestsubsets}}{
An object of class \code{LRCbestsubsets_pred} (which inherits
from \code{data.frame})
that contains the predicted class for each observation.  The columns indicated
by \code{truthCol} and \code{keepCols} are included if they were requested.}
}
\description{
Predict (or classify) new data using a fitted best subsets logistic regression
classifier
}
\author{
Landon Sego
}
\seealso{
See \code{\link{LRCbestsubsets}} for examples
}

