% Generated by roxygen2 (4.0.1): do not edit by hand
\name{predict.LRCglmnet}
\alias{predict.LRCglmnet}
\title{Predict (or classify) new data using a fitted glmnet logistic regression classifier}
\usage{
\method{predict}{LRCglmnet}(LRCglmnet_object, newdata, truthCol = NULL,
  keepCols = NULL)
}
\arguments{
\item{LRCglmnet_object}{An object of class \code{LRCglmnet}, returned by
\code{\link{LRCglmnet}},
which contains the optimally-trained elastic net logistic regression classifier}

\item{newdata}{A dataframe or matrix containing the new set of observations to
be predicted, as well as an optional column of true labels.
\code{newdata} must contain all of the column names that were used
to fit the elastic net logistic regression classifier.}

\item{truthCol}{The column number or column name in \code{newdata} that contains the
true labels. Optional.}

\item{keepCols}{A numeric vector of column numbers (or a character vector of
column names) that will be 'kept' and returned with the predictions. Optional.}
}
\value{
An object of class \code{LRCpred} (which inherits
from \code{data.frame})
that contains the predicted class for each observation.  The columns indicated
by \code{truthCol} and \code{keepCols} are included if they were requested.
}
\description{
Predict (or classify) new data using a fitted glmnet logistic regression classifier
}
\author{
Landon Sego
}
\seealso{
See \code{\link{LRCglmnet}} for examples
}

