% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.permuteTest.R
\name{print.permuteTest}
\alias{print.permuteTest}
\title{Print the p-value of the permutation test}
\usage{
\method{print}{permuteTest}(x, ...)
}
\arguments{
\item{x}{an object of class \code{permuteTest} returned by \code{\link{permuteTest.glmnetLRC}}.}

\item{\dots}{Arguments passed to \code{\link{print.default}}.}
}
\description{
Print the p-value of the permutation test
}
\seealso{
See \code{\link{glmnetLRC}} for an example.
}
\author{
Landon Sego
}
