% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testdata.R
\docType{data}
\name{testdata}
\alias{testdata}
\title{Test data set from the Vorbitrap mass spectrometry instrument}
\format{A dataframe with 100 observations and 96 variables}
\source{
\url{http://pubs.acs.org/doi/suppl/10.1021/pr401143e}
}
\description{
Test data set from the Vorbitrap mass spectrometry instrument
}

