% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traindata.R
\docType{data}
\name{traindata}
\alias{traindata}
\title{Training data set from the Vorbitrap mass spectrometry instrument}
\format{A dataframe with 467 observations and 99 variables}
\description{
Training data set from the Vorbitrap mass spectrometry instrument
}
\examples{
data(traindata)
str(traindata)
}
\references{
Amidan BG, Orton DJ, LaMarche BL, Monroe ME, Moore RJ,
Venzin AM, Smith RD, Sego LH, Tardiff MF, Payne SH. 2014.
Signatures for Mass Spectrometry Data Quality.
Journal of Proteome Research. 13(4), 2215-2222.
\url{http://pubs.acs.org/doi/abs/10.1021/pr401143e}
}

