% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{extract}
\alias{extract}
\alias{missingpreds}
\alias{permuteTest}
\title{Generics for \code{extract}, \code{missingpreds}, and \code{permuteTest} methods}
\usage{
extract(object, ...)
missingpreds(object, ...)
permuteTest(object, ...)
}
\arguments{
\item{object}{The object on which the generic operates}

\item{\dots}{Arguments passed to specific methods}
}
\description{
Generics for \code{extract}, \code{missingpreds}, and \code{permuteTest} methods
}
