% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnetLRC-package.R
\docType{package}
\name{glmnetLRC-package}
\alias{glmnetLRC-package}
\title{Lasso and elastic-net logistic regression classification with an arbitrary loss function}
\description{
Lasso and elastic-net logistic regression classification (LRC) with an arbitrary loss function
}
\details{
\pkg{lrc} extends the \pkg{glmnet} package for training elastic net LRCs with a user-specified
discrete loss function.

There are a handful of functions you'll need (along with their associated methods):
\code{\link{lossMatrix}}, \code{\link{glmnetLRC}}, 
and \code{\link{predict.glmnetLRC}}
}
\author{
Landon Sego, Alex Venzin, John Ramey
}

