% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{Mojave data}
\alias{Mojave data}
\title{Environmental data from the Mojave desert}
\format{A dataframe with 633 obs and 11 variables, \code{cheatGrass} is the response
variable, the rest are location or predictor variables}
\description{
Environmental data from the Mojave desert
}
\examples{
data(Mojave)
head(Mojave)
str(Mojave)
}

