% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.LRCpred.R
\name{summary.LRCpred}
\alias{summary.LRCpred}
\title{Summarize predictions of logistic regression classifier}
\usage{
\method{summary}{LRCpred}(object, ...)
}
\arguments{
\item{object}{an object of class \code{LRCpred} returned by
\code{\link{predict.glmnetLRC}}.}

\item{\dots}{Arguments passed to \code{\link{print}} methods}
}
\value{
Prints and invisibly returns the following information:
\itemize{

\item If \code{truthCol} was provided in the call to
\code{\link{predict.glmnetLRC}}, the sensitivity, specificity, false negative rate,
false positive rate, and accuracy for the class designated by the second level of
the \code{truthLabels} argument
in \code{\link{glmnetLRC}} are calculated.  A summary of the predicted probabilities, according to the true class, is
also provided.

\item If \code{truthCol = NULL} in the call to \code{\link{predict.glmnetLRC}}
a tabulation of the number of predictions for each class is shown, along with a summary of the probabilities according
to each predicted class.
}
}
\description{
Summarize the predicted probabilities of the classifier, and, if possible,
calculate accuracy, sensitivity, specificity, false positive rate, and false
negative rate.
}
\author{
Landon Sego
}
\seealso{
See \code{\link{glmnetLRC}} and \code{\link{glmnetLRC_fit}}
for examples.
}

