% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.LRCpred.R
\name{summary.LRCpred}
\alias{summary.LRCpred}
\title{Calculate performance metrics of logistic regression classifiers}
\usage{
\method{summary}{LRCpred}(LRCpredObject)
}
\arguments{
\item{LRCpredObject}{an object of class \code{LRC_pred} returned by
\code{\link{predict.LRCbestsubsets}} or \code{\link{predict.LRCglmnet}}.}
}
\value{
\itemize{
\item If \code{truthCol} was provided in the call to
\code{\link{predict.LRCglmnet}} or \code{\link{predict.LRCbestsubsets}}, a
\code{data.frame} is returned with the sensitivity, specificity, false negative rate,
false positive rate, and accuracy for the class designated by the second level of
the \code{truthLabels} argument in \code{\link{LRCglmnet}} or
\code{\link{LRCbestsubsets}}.

\item If \code{truthCol = NULL} in the call to \code{\link{predict.LRCglmnet}},
or \code{\link{predict.LRCbestsubsets}}, a tabulation
of the number of predictions for each class is returned.
}
}
\description{
Calculate accuracy, sensitivity, specificity, false positive rate, and false
negative rate for predictions from logistic regression classifiers
}
\author{
Landon Sego
}

