% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.LRCpred.R
\name{summary.LRCpred}
\alias{print.summaryLRCpred}
\alias{summary.LRCpred}
\title{Summarize predictions of logistic regression classifier}
\usage{
\method{summary}{LRCpred}(object, ...)

\method{print}{summaryLRCpred}(x, ...)
}
\arguments{
\item{object}{An object of class \code{LRCpred} returned by
\code{\link{predict.glmnetLRC}}.}

\item{x}{An object of class \code{summaryLRCpred}.}

\item{\dots}{Arguments passed to \code{\link{print}} methods. Ignored by the \code{summary} method.}
}
\value{
Returns a \code{summaryLRCpred} object.  If \code{truthCol} was provided in the call to
\code{\link{predict.glmnetLRC}}, the result is a list with the following elements:

\describe{
\item{ConfusionMatrixMetrics}{A matrix with the sensitivity, specificity, false negative rate,
false positive rate, and accuracy for the class designated by the second level of
the \code{truthLabels} argument provided to \code{\link{glmnetLRC}}}
\item{PredProbSummary}{A numeric summary of the predicted probabilities, according to the true class}
}

If \code{truthCol} was not provided in the call to \code{\link{predict.glmnetLRC}}, the result is a list with
the following elements:
\describe{
\item{PredClassSummary}{A tabulation of the number of predictions in each class}
\item{PredProbSummary}{A numeric summary of the predicted probabilities, according to the predicted class}
}
}
\description{
Summarize the predicted probabilities of the classifier, and, if possible,
calculate accuracy, sensitivity, specificity, false positive rate, and false
negative rate.
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: Prints a \code{summaryLRCpred} object in a readable format.
}}
\author{
Landon Sego
}
\seealso{
See \code{\link{glmnetLRC}} for examples.
}

