% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{traindata}
\alias{traindata}
\title{Training data set from the Vorbitrap mass spectrometry instrument}
\format{A dataframe with 467 observations and 99 variables}
\source{
\url{http://pubs.acs.org/doi/suppl/10.1021/pr401143e}
}
\description{
Training data set from the Vorbitrap mass spectrometry instrument
}

